requirements.txt
Flask==2.2.3
pandas==1.5.2
openpyxl==3.0.10  # Needed for reading .xlsx files
requests==2.28.1
python-dotenv==0.21.0
groq==0.2.1  # Replace with the actual version if needed
